oasp4j-jpa compatiblity for DAO with paging support
