+++ boms/minimal/pom.xml
@@ -132,6 +132,11 @@
       </dependency>
       <dependency>
         <groupId>io.oasp.java.modules</groupId>
+        <artifactId>oasp4j-jpa-dao-envers</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.oasp.java.modules</groupId>
         <artifactId>oasp4j-web</artifactId>
         <version>${project.version}</version>
       </dependency>
