+++ modules/jpa-basic/pom.xml
@@ -15,7 +15,7 @@
   <description>JPA-based persistence infrastructure of the Open Application Standard Platform for Java (OASP4J).</description>
 
   <dependencies>
-   <dependency>
+    <dependency>
       <groupId>io.oasp.java.modules</groupId>
       <artifactId>oasp4j-basic</artifactId>
     </dependency>
