+++ modules/jpa-dao-envers/src/main/java/io/oasp/module/jpa/dataaccess/api/AdvancedRevisionListener.java
