+++ modules/jpa-dao-envers/src/main/java/io/oasp/module/jpa/dataaccess/base/dao/AbstractRevisionedDaoImpl.java
@@ -1,0 +1,14 @@
+package io.oasp.module.jpa.dataaccess.base.dao;
+
+import io.oasp.module.jpa.dataaccess.api.MutablePersistenceEntity;
+import io.oasp.module.jpa.dataaccess.api.dao.IRevisionedDao;
+
+/**
+ * Abstract base implementation of {@link IRevisionedDao} interface.
+ *
+ * @param <ENTITY> is the type of the managed {@link MutablePersistenceEntity entity}.
+ */
+public abstract class AbstractRevisionedDaoImpl<ENTITY extends MutablePersistenceEntity<Long>> extends
+    AbstractGenericRevisionedDaoImpl<Long, ENTITY> implements IRevisionedDao<ENTITY> {
+
+}
