+++ modules/jpa-dao-envers/src/main/java/io/oasp/module/jpa/dataaccess/impl/LazyRevisionMetadata.java
