+++ modules/jpa-dao/src/test/java/io/oasp/example/component/dataaccess/impl/dao/BarDaoImpl.java
@@ -4,13 +4,13 @@ import javax.inject.Named;
 
 import io.oasp.example.component.dataaccess.api.BarEntity;
 import io.oasp.example.component.dataaccess.api.dao.BarDao;
-import io.oasp.module.jpa.dataaccess.base.AbstractDao;
+import io.oasp.module.jpa.dataaccess.base.dao.AbstractDaoImpl;
 
 /**
  * Implementation of {@link BarDao}.
  */
 @Named
-public class BarDaoImpl extends AbstractDao<BarEntity> implements BarDao {
+public class BarDaoImpl extends AbstractDaoImpl<BarEntity> implements BarDao {
 
   @Override
   protected Class<BarEntity> getEntityClass() {
