+++ modules/jpa-dao/src/test/java/io/oasp/module/jpa/dataaccess/base/AbstractGenericDaoTest.java
@@ -9,11 +9,11 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import io.oasp.example.TestApplication;
 import io.oasp.example.component.dataaccess.api.BarEntity;
 import io.oasp.example.component.dataaccess.impl.BarDaoTxBean;
-import io.oasp.module.jpa.dataaccess.api.GenericDao;
+import io.oasp.module.jpa.dataaccess.api.dao.IGenericDao;
 import io.oasp.module.test.common.base.ComponentTest;
 
 /**
- * Test class to test the {@link GenericDao}.
+ * Test class to test the {@link IGenericDao}.
  */
 @SpringBootTest(classes = { TestApplication.class }, webEnvironment = WebEnvironment.NONE)
 public class AbstractGenericDaoTest extends ComponentTest {
@@ -22,7 +22,7 @@ public class AbstractGenericDaoTest extends ComponentTest {
   private BarDaoTxBean testBean;
 
   /**
-   * Test of {@link GenericDao#forceIncrementModificationCounter(Object)}. Ensures that the modification counter is
+   * Test of {@link IGenericDao#forceIncrementModificationCounter(Object)}. Ensures that the modification counter is
    * updated after the call of that method when the transaction is closed.
    */
   @Test
