+++ modules/jpa-envers/src/main/java/io/oasp/module/jpa/dataaccess/base/AbstractRevisionedDao.java
@@ -1,14 +1,15 @@
 package io.oasp.module.jpa.dataaccess.base;
 
-import io.oasp.module.jpa.dataaccess.api.RevisionedDao;
 import io.oasp.module.jpa.dataaccess.api.MutablePersistenceEntity;
+import io.oasp.module.jpa.dataaccess.api.RevisionedDao;
 
 /**
  * Abstract base implementation of {@link RevisionedDao} interface.
  *
  * @param <ENTITY> is the type of the managed {@link MutablePersistenceEntity entity}.
  */
-public abstract class AbstractRevisionedDao<ENTITY extends MutablePersistenceEntity<Long>> extends
-    AbstractGenericRevisionedDao<Long, ENTITY> implements RevisionedDao<ENTITY> {
+@Deprecated
+public abstract class AbstractRevisionedDao<ENTITY extends MutablePersistenceEntity<Long>>
+    extends AbstractGenericRevisionedDao<Long, ENTITY> implements RevisionedDao<ENTITY> {
 
 }
