+++ modules/jpa-spring-data/pom.xml
@@ -23,6 +23,10 @@
       <artifactId>oasp4j-jpa-basic</artifactId>
     </dependency>
     <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>oasp4j-jpa-dao-envers</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.springframework.data</groupId>
       <artifactId>spring-data-jpa</artifactId>
     </dependency>
