+++ modules/jpa/src/main/java/io/oasp/module/jpa/dataaccess/api/Dao.java
@@ -10,6 +10,7 @@ import net.sf.mmm.util.entity.api.PersistenceEntity;
  *
  * @param <E> is the generic type of the {@link PersistenceEntity}.
  */
+@Deprecated
 public interface Dao<E extends PersistenceEntity<Long>> extends GenericDao<Long, E> {
 
 }
