+++ modules/jpa/src/main/java/io/oasp/module/jpa/dataaccess/base/AbstractDao.java
@@ -1,4 +1,4 @@
-package io.oasp.module.jpa.common.base;
+package io.oasp.module.jpa.dataaccess.base;
 
 import net.sf.mmm.util.entity.api.PersistenceEntity;
 
@@ -7,11 +7,8 @@ import io.oasp.module.jpa.dataaccess.api.Dao;
 /**
  * Abstract base implementation of {@link Dao} interface.
  *
- * @param <E> is the generic type of the {@link PersistenceEntity}. Should be derived from
- *        {@link io.oasp.module.jpa.dataaccess.api.AbstractPersistenceEntity}.
+ * @param <E> is the generic type of the {@link PersistenceEntity}.
  *
- * @deprecated It is only provided to support old functionality of DAO, instead use
- *             io.oasp.module.jpa.dataaccess.base.AbstractDao instead
  */
 @Deprecated
 public abstract class AbstractDao<E extends PersistenceEntity<Long>> extends AbstractGenericDao<Long, E>
