+++ modules/jpa/src/main/java/io/oasp/module/jpa/dataaccess/base/AbstractMasterDataDao.java
@@ -11,6 +11,7 @@ import io.oasp.module.jpa.dataaccess.api.MasterDataDao;
  *
  * @param <E> is the generic type of the {@link PersistenceEntity}.
  */
+@Deprecated
 public abstract class AbstractMasterDataDao<E extends PersistenceEntity<Long>> extends AbstractDao<E>
     implements MasterDataDao<E> {
